add an API to create belief structures.
This will ease the parsing of new types of files and will add the
possibility to create structure on the fly later.

Currently the function do not really check for consistency in the
belief structure, meaning that you can add a sensor model which
contains beliefs about another frame of discernment for instance.