Added an xml parser for the configuration.
The xml format will allow to have one file containing everything needed
for a model. Currently there is no verification via xml schema or DTD.

The new parser is compiled as another library depending on THEGAME in
order to avoid adding dependencies to the core part of the library.
Someone who wants to use xml with THEGAME will have to link to
THEGAME-xml.